@@ -1,44 +1,28 @@
-/*
- * DocumentType.java
- *
- * Copyright (c) 2011, Database Research Group, Institute of Computer Science, University of Heidelberg.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the GNU General Public License.
- *
- * authors: Andreas Fay, Jannik Strötgen
- * email:  fay@stud.uni-heidelberg.de, stroetgen@uni-hd.de
- *
- * HeidelTime is a multilingual, cross-domain temporal tagger.
- * For details, see http://dbs.ifi.uni-heidelberg.de/heideltime
- */
-
 package de.unihd.dbs.heideltime.standalone;
 
 /**
- * Type of document to be processed by HeidelTime
- * 
- * @author Andreas Fay, University of Heidelberg
- * @version 1.0
+ * Legacy constants to transition to the enum at {@link de.unihd.dbs.uima.annotator.heideltime.DocumentType}.
+ *
+ * Because we cannot subclass enums, this will not be binary compatible,
+ * but at least we get compile time compatibility.
+ *
+ * @author Erich Schubert
  */
-public enum DocumentType {
-	NARRATIVES {
-		public String toString() {
-			return "narratives";
-		}
-	},
-	NEWS {
-		public String toString() {
-			return "news";
-		}
-	},
-	COLLOQUIAL {
-		public String toString() {
-			return "colloquial";
-		}
-	},
-	SCIENTIFIC {
-		public String toString() {
-			return "scientific";
-		}
-	}
+@Deprecated
+public final class DocumentType {
+	/** Use {@link de.unihd.dbs.uima.annotator.heideltime.DocumentType.NARRATIVE} instead. */
+	@Deprecated
+	public static final de.unihd.dbs.uima.annotator.heideltime.DocumentType NARRATIVES = de.unihd.dbs.uima.annotator.heideltime.DocumentType.NARRATIVE;
+
+	/** Use {@link de.unihd.dbs.uima.annotator.heideltime.DocumentType.NEWS} instead. */
+	@Deprecated
+	public static final de.unihd.dbs.uima.annotator.heideltime.DocumentType NEWS = de.unihd.dbs.uima.annotator.heideltime.DocumentType.NEWS;
+
+	/** Use {@link de.unihd.dbs.uima.annotator.heideltime.DocumentType.COLLOQUIAL} instead. */
+	@Deprecated
+	public static final de.unihd.dbs.uima.annotator.heideltime.DocumentType COLLOQUIAL = de.unihd.dbs.uima.annotator.heideltime.DocumentType.COLLOQUIAL;
+
+	/** Use {@link de.unihd.dbs.uima.annotator.heideltime.DocumentType.SCIENTIFIC} instead. */
+	@Deprecated
+	public static final de.unihd.dbs.uima.annotator.heideltime.DocumentType SCIENTIFIC = de.unihd.dbs.uima.annotator.heideltime.DocumentType.SCIENTIFIC;
 }
