@@ -20,7 +20,7 @@ import org.apache.uima.resource.impl.ConfigurationManager_impl;
 import org.apache.uima.resource.impl.ResourceManager_impl;
 
 import de.unihd.dbs.heideltime.standalone.Config;
-import de.unihd.dbs.heideltime.standalone.DocumentType;
+import de.unihd.dbs.uima.annotator.heideltime.DocumentType;
 import de.unihd.dbs.uima.annotator.heideltime.resources.Language;
 
 /**
