@@ -15,9 +15,13 @@ import java.util.jar.JarFile;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import de.unihd.dbs.uima.annotator.heideltime.utilities.Logger;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 public class ResourceScanner {
+	/** Class logger */
+	private static final Logger LOG = LoggerFactory.getLogger(ResourceScanner.class);
+	
 	private static ResourceScanner INSTANCE = null;
 
 	/**
@@ -100,12 +104,10 @@ public class ResourceScanner {
 	
 
 	public static void main(String[] args) {
-		@SuppressWarnings("unused")
-		ResourceScanner rs = null;
 		try {
-			rs = new ResourceScanner();
+			new ResourceScanner();
 		} catch (Exception e) {
-			e.printStackTrace();
+			LOG.error(e.getMessage(), e);
 		}
 	}
 
@@ -125,7 +127,7 @@ public class ResourceScanner {
 			Pattern rulePattern = Pattern.compile(language + "/rules/resources_rules_(.+)\\.txt$");
 			
 			if (entry.getValue().isDirectory()) {
-				Logger.printDetail(ResourceScanner.class, "Testing " + entry.getKey());
+				LOG.trace("Testing {}", entry.getKey());
 				/*
 				 * our conditions for something being a resources folder: the resource
 				 * folder must contain at least the following folders:
@@ -136,9 +138,9 @@ public class ResourceScanner {
 				 * |- rules
 				 */
 				
-				Boolean repatternExists = false;
-				Boolean normalizationExists = false;
-				Boolean ruleExists = false;
+				boolean repatternExists = false;
+				boolean normalizationExists = false;
+				boolean ruleExists = false;
 		
 				for(String entryName : jarContents.keySet()) {
 					if(!repatternExists && repatternPattern.matcher(entryName).matches()) {
@@ -153,11 +155,11 @@ public class ResourceScanner {
 				}
 				
 				if(!repatternExists || !normalizationExists || !ruleExists) {
-					Logger.printDetail(ResourceScanner.class, "We need at least one readable resource file of each type to run.");
+					LOG.debug("We need at least one readable resource file of each type to run (in {})", entry.getKey());
 					continue;
 				}
 		
-				Logger.printDetail(ResourceScanner.class, "Valid resource folder.");
+				LOG.trace("Valid resource folder: {}", entry.getKey());
 
 				// at this point, the folder is obviously a language resource folder => collect streams
 				this.repatterns.put(language, new ResourceMap());
@@ -191,10 +193,10 @@ public class ResourceScanner {
 		for (File supposedLanguagePath : pathContents) {
 			String language = supposedLanguagePath.getName();
 			if (supposedLanguagePath.isDirectory()) {
-				Logger.printDetail(ResourceScanner.class, "Testing " + supposedLanguagePath.getAbsolutePath());
+				LOG.trace("Testing {}", supposedLanguagePath);
 		
 				if (!supposedLanguagePath.exists()) {
-					Logger.printDetail(ResourceScanner.class, "This path doesn't exist.");
+					LOG.debug("This path doesn't exist.");
 					continue;
 				}
 		
@@ -215,8 +217,7 @@ public class ResourceScanner {
 				if (!repatternFolder.exists() || !repatternFolder.canRead() || !repatternFolder.isDirectory()
 						|| !normalizationFolder.exists() || !normalizationFolder.canRead() || !normalizationFolder.isDirectory() 
 						|| !ruleFolder.exists() || !ruleFolder.canRead() || !ruleFolder.isDirectory()) {
-					Logger.printDetail(ResourceScanner.class, "We need at least the folders repattern, normalization and rules in this folder.");
-					
+					LOG.debug("We need at least the folders repattern, normalization and rules in folder '{}'.", supposedLanguagePath);					
 					continue;
 				}
 		
@@ -244,11 +245,11 @@ public class ResourceScanner {
 						|| !repatternFiles[0].exists() || !repatternFiles[0].canRead() || !repatternFiles[0].isFile()
 						|| !normalizationFiles[0].exists() || !normalizationFiles[0].canRead() || !normalizationFiles[0].isFile()
 						|| !ruleFiles[0].exists() || !ruleFiles[0].canRead() || !ruleFiles[0].isFile()) {
-					Logger.printDetail(ResourceScanner.class, "We need at least one readable resource file of each type to run.");
+					LOG.debug("We need at least one readable resource file of each type to run in '{}'", supposedLanguagePath);
 					continue;
 				}
 		
-				Logger.printDetail(ResourceScanner.class, "Valid resource folder.");
+				LOG.trace("Valid resource folder: {}", supposedLanguagePath);
 
 				// at this point, the folder is obviously a language resource folder => collect streams
 				this.repatterns.put(language, new ResourceMap());
