@@ -16,8 +16,11 @@ import org.apache.uima.analysis_component.JCasAnnotator_ImplBase;
 import org.apache.uima.analysis_engine.AnalysisEngineProcessException;
 import org.apache.uima.cas.FSIterator;
 import org.apache.uima.jcas.JCas;
+import org.apache.uima.jcas.tcas.Annotation;
 
-import de.unihd.dbs.uima.annotator.heideltime.utilities.Logger;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import de.unihd.dbs.uima.types.heideltime.Sentence;
 import de.unihd.dbs.uima.types.heideltime.Token;
 
@@ -34,7 +37,8 @@ import edu.stanford.nlp.process.PTBTokenizer.PTBTokenizerFactory;
  *
  */
 public class StanfordPOSTaggerWrapper extends JCasAnnotator_ImplBase {
-	private Class<?> component = this.getClass();
+	/** Class logger */
+	private static final Logger LOG = LoggerFactory.getLogger(StanfordPOSTaggerWrapper.class);
 	
 	// definitions of what names these parameters have in the wrapper's descriptor file
 	public static final String PARAM_MODEL_PATH = "model_path";
@@ -66,7 +70,7 @@ public class StanfordPOSTaggerWrapper extends JCasAnnotator_ImplBase {
 
 		// check if the model file exists
 		if(model_path == null) {
-			Logger.printError(component, "The model file for the Stanford Tagger was not correctly specified.");
+			LOG.error("The model file for the Stanford Tagger was not correctly specified.");
 			System.exit(-1);
 		}
 		
@@ -81,8 +85,7 @@ public class StanfordPOSTaggerWrapper extends JCasAnnotator_ImplBase {
 				mt = new MaxentTagger(model_path, new TaggerConfig("-model", model_path), false);
 			}
 		} catch(Exception e) {
-			e.printStackTrace();
-			Logger.printError(component, "MaxentTagger could not be instantiated with the supplied model("+model_path+") and config("+config_path+") file.");
+			LOG.error("MaxentTagger could not be instantiated with the supplied model("+model_path+") and config("+config_path+") file.", e);
 			System.exit(-1);
 		}
 	}
@@ -91,7 +94,7 @@ public class StanfordPOSTaggerWrapper extends JCasAnnotator_ImplBase {
 	 * Method that gets called to process the documents' cas objects
 	 */
 	public void process(JCas jcas) throws AnalysisEngineProcessException {
-		Integer offset = 0; // a cursor of sorts to keep up with the position in the document text
+		int offset = 0; // a cursor of sorts to keep up with the position in the document text
 		
 		// grab the document text
 		String docText = jcas.getDocumentText();
@@ -100,6 +103,7 @@ public class StanfordPOSTaggerWrapper extends JCasAnnotator_ImplBase {
 		fac.setOptions("ptb3Escaping=false,untokenizable=noneKeep");
 		List<List<HasWord>> tokenArray = MaxentTagger.tokenizeText(new StringReader(docText), fac);
 		
+		int sentences = 0;
 		// iterate over sentences in this document
 		for(List<HasWord> sentenceToken : tokenArray) {
 			List<TaggedWord> taggedSentence = mt.tagSentence(sentenceToken);
@@ -107,9 +111,10 @@ public class StanfordPOSTaggerWrapper extends JCasAnnotator_ImplBase {
 			
 			// create a sentence object. gets added to index or discarded depending on configuration
 			Sentence sentence = new Sentence(jcas);
+			sentence.setSentenceId(++sentences);
 			sentence.setBegin(offset);
 			
-			Integer wordCount = 0;
+			int wordCount = 0;
 			// iterate over words in this sentence
 			for(HasWord wordToken : sentenceToken) {
 				Token t = new Token(jcas);
@@ -123,8 +128,8 @@ public class StanfordPOSTaggerWrapper extends JCasAnnotator_ImplBase {
 				String thisWord = wordToken.word();
 				
 				if(docText.indexOf(thisWord, offset) < 0) {
-					Logger.printDetail(component, "A previously tagged token wasn't found in the document text: \"" + thisWord + "\". " +
-							"This may be due to unpredictable punctuation tokenization; hence this token isn't tagged.");
+					LOG.debug("A previously tagged token wasn't found in the document text: \"{}\". " +
+							"This may be due to unpredictable punctuation tokenization; hence this token isn't tagged.", thisWord);
 					continue; // jump to next token: discards token
 				} else {
 					offset = docText.indexOf(thisWord, offset); // set cursor to the starting position of token in docText
@@ -152,7 +157,7 @@ public class StanfordPOSTaggerWrapper extends JCasAnnotator_ImplBase {
 		}
 		
 		// TODO: DEBUG
-		FSIterator fsi = jcas.getAnnotationIndex(Sentence.type).iterator();
+		FSIterator<Annotation> fsi = jcas.getAnnotationIndex(Sentence.type).iterator();
 		while(fsi.hasNext()) {
 			Sentence s = (Sentence) fsi.next();
 			if(s.getBegin() < 0 || s.getEnd() < 0) {
@@ -161,7 +166,7 @@ public class StanfordPOSTaggerWrapper extends JCasAnnotator_ImplBase {
 				System.exit(-1);
 			}
 		}
-		FSIterator fsi2 = jcas.getAnnotationIndex(Token.type).iterator();
+		FSIterator<Annotation> fsi2 = jcas.getAnnotationIndex(Token.type).iterator();
 		while(fsi2.hasNext()) {
 			Token t = (Token) fsi2.next();
 			if(t.getBegin() < 0 || t.getEnd() < 0) {
