@@ -23,15 +23,17 @@ import org.apache.uima.collection.CasConsumer_ImplBase;
 import org.apache.uima.jcas.JCas;
 import org.apache.uima.resource.ResourceInitializationException;
 import org.apache.uima.resource.ResourceProcessException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
 
-import de.unihd.dbs.uima.annotator.heideltime.utilities.Logger;
 import de.unihd.dbs.uima.types.heideltime.Dct;
 import de.unihd.dbs.uima.types.heideltime.Timex3;
 
 public class TempEval3Writer extends CasConsumer_ImplBase {
-	private Class<?> component = this.getClass();
+	/** Class logger */
+	private static final Logger LOG = LoggerFactory.getLogger(TempEval3Writer.class);
 
 	private static final String PARAM_OUTPUTDIR = "OutputDir";
 	
@@ -45,13 +47,13 @@ public class TempEval3Writer extends CasConsumer_ImplBase {
 		
 		if (!mOutputDir.exists()) {
 			if(!mOutputDir.mkdirs()) {
-				Logger.printError(component, "Couldn't create non-existant folder "+mOutputDir.getAbsolutePath());
+				LOG.error("Couldn't create non-existant folder "+mOutputDir.getAbsolutePath());
 				throw new ResourceInitializationException();
 			}
 		}
 		
 		if(!mOutputDir.canWrite()) {
-			Logger.printError(component, "Folder "+mOutputDir.getAbsolutePath()+" is not writable.");
+			LOG.error("Folder "+mOutputDir.getAbsolutePath()+" is not writable.");
 			throw new ResourceInitializationException();
 		}
 	}
@@ -99,8 +101,7 @@ public class TempEval3Writer extends CasConsumer_ImplBase {
 			db = dbf.newDocumentBuilder();
 			doc = db.newDocument();
 		} catch (ParserConfigurationException e) {
-			e.printStackTrace();
-			Logger.printError(component, "XML Builder could not be instantiated");
+			LOG.error("XML Builder could not be instantiated", e);
 		}
 		
 		// create the TimeML root element
@@ -166,7 +167,7 @@ public class TempEval3Writer extends CasConsumer_ImplBase {
 						prevT = thisT; // this iteration's prevT was removed; setting for new iteration 
 					}
 					
-					Logger.printError(component, "Two overlapping Timexes have been discovered:" + System.getProperty("line.separator")
+					LOG.error("Two overlapping Timexes have been discovered:" + System.getProperty("line.separator")
 							+ "Timex A: " + prevT.getCoveredText() + " [\"" + prevT.getTimexValue() + "\" / " + prevT.getBegin() + ":" + prevT.getEnd() + "]" 
 							+ System.getProperty("line.separator")
 							+ "Timex B: " + removedT.getCoveredText() + " [\"" + removedT.getTimexValue() + "\" / " + removedT.getBegin() + ":" + removedT.getEnd() + "]" 
@@ -253,17 +254,14 @@ public class TempEval3Writer extends CasConsumer_ImplBase {
 			// transform
 			transformer.transform(source, result);
 		} catch (IOException e) { // something went wrong with the bufferedwriter
-			e.printStackTrace();
-			Logger.printError(component, "File "+outFile.getAbsolutePath()+" could not be written.");
+			LOG.error("File "+outFile.getAbsolutePath()+" could not be written.", e);
 		} catch (TransformerException e) { // the transformer malfunctioned (call optimus prime)
-			e.printStackTrace();
-			Logger.printError(component, "XML transformer could not be properly initialized.");
+			LOG.error("XML transformer could not be properly initialized.", e);
 		} finally { // clean up for the bufferedwriter
 			try {
 				bw.close();
 			} catch(IOException e) {
-				e.printStackTrace();
-				Logger.printError(component, "File "+outFile.getAbsolutePath()+" could not be closed.");
+				LOG.error("File "+outFile.getAbsolutePath()+" could not be closed.", e);
 			}
 		}
 	}
